+++ R/init.R
@@ -262,7 +262,6 @@ renv_init_validate_project <- function(project, force) {
 
 renv_init_settings <- function(project, settings) {
 
-  ensure_directory(file.path(project, "renv"))
   defaults <- renv_settings_get(project)
   merged <- renv_settings_merge(defaults, settings)
   renv_settings_persist(project, merged)