+++ tests/testthat/test-install.R
@@ -1,3 +1,4 @@
+
 context("Install")
 
 test_that("install works when DESCRIPTION contains no dependencies", {
@@ -360,8 +361,10 @@ test_that("issue #609", {
 })
 
 test_that("we can install packages from git remotes within subdirs", {
+
   skip_on_cran()
   skip_on_ci()
+  skip("unreliable test")
 
   renv_tests_scope("subdir")